+++ packages/audit/package.json
@@ -13,7 +13,7 @@
   "scripts": {
     "lint": "tslint -c tslint.json src/**/*.ts test/**/*.ts",
     "tsc": "rimraf lib && tsc",
-    "test": "pnpm run tsc && pnpm run lint",
+    "test": "pnpm run tsc && pnpm run lint && ts-node test",
     "prepublishOnly": "pnpm run tsc"
   },
   "repository": "https://github.com/pnpm/pnpm/blob/master/packages/audit",
