+++ packages/audit/test/lockfileToAuditTree.ts
@@ -1,0 +1,1 @@
+import '../lib/'
